+++ src/domain/datasets/domain/src/services/mod.rs
@@ -8,5 +8,7 @@
 // by the Apache License, Version 2.0.
 
 mod dataset_env_var_service;
+mod dataset_key_value_service;
 
 pub use dataset_env_var_service::*;
+pub use dataset_key_value_service::*;