+++ src/infra/core/tests/tests/ingest/test_fetch.rs
@@ -8,6 +8,7 @@
 // by the Apache License, Version 2.0.
 
 use std::assert_matches::assert_matches;
+use std::collections::HashMap;
 use std::sync::{Arc, Mutex};
 
 use chrono::prelude::*;
@@ -17,7 +18,7 @@ use indoc::indoc;
 use kamu::domain::*;
 use kamu::ingest::*;
 use kamu::utils::docker_images::BUSYBOX;
-use kamu_datasets_services::DatasetEnvVarServiceStaticImpl;
+use kamu_datasets_services::DatasetKeyValueServiceStaticImpl;
 use opendatafabric::*;
 use url::Url;
 
@@ -52,6 +53,7 @@ async fn test_fetch_url_file() {
                 None,
                 &target_path,
                 &Utc::now(),
+                &HashMap::new(),
                 None
             )
             .await,
@@ -71,6 +73,7 @@ async fn test_fetch_url_file() {
             None,
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await
@@ -91,6 +94,7 @@ async fn test_fetch_url_file() {
             update.source_state.as_ref(),
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await
@@ -108,6 +112,7 @@ async fn test_fetch_url_file() {
             update.source_state.as_ref(),
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await
@@ -141,6 +146,7 @@ async fn test_fetch_url_http_unreachable() {
                 None,
                 &target_path,
                 &Utc::now(),
+                &HashMap::new(),
                 None
             )
             .await,
@@ -175,6 +181,7 @@ async fn test_fetch_url_http_not_found() {
                 None,
                 &target_path,
                 &Utc::now(),
+                &HashMap::new(),
                 None
             )
             .await,
@@ -216,6 +223,7 @@ async fn test_fetch_url_http_ok() {
             None,
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             Some(listener.clone()),
         )
         .await
@@ -249,6 +257,7 @@ async fn test_fetch_url_http_ok() {
             update.source_state.as_ref(),
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await
@@ -267,6 +276,7 @@ async fn test_fetch_url_http_ok() {
             update.source_state.as_ref(),
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await
@@ -286,6 +296,7 @@ async fn test_fetch_url_http_ok() {
                 update.source_state.as_ref(),
                 &target_path,
                 &Utc::now(),
+                &HashMap::new(),
                 None
             )
             .await,
@@ -332,6 +343,7 @@ async fn test_fetch_url_http_env_interpolation() {
                 None,
                 &target_path,
                 &Utc::now(),
+                &HashMap::new(),
                 Some(listener.clone())
             )
             .await,
@@ -349,6 +361,7 @@ async fn test_fetch_url_http_env_interpolation() {
             None,
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             Some(listener.clone()),
         )
         .await
@@ -407,6 +420,7 @@ async fn test_fetch_url_ftp_ok() {
             None,
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             Some(listener.clone()),
         )
         .await
@@ -465,6 +479,7 @@ async fn test_fetch_files_glob() {
                 None,
                 &target_path,
                 &Utc::now(),
+                &HashMap::new(),
                 None
             )
             .await,
@@ -484,6 +499,7 @@ async fn test_fetch_files_glob() {
             None,
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await
@@ -512,6 +528,7 @@ async fn test_fetch_files_glob() {
             update.source_state.as_ref(),
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await
@@ -529,6 +546,7 @@ async fn test_fetch_files_glob() {
             update.source_state.as_ref(),
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await
@@ -549,6 +567,7 @@ async fn test_fetch_files_glob() {
             update.source_state.as_ref(),
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await
@@ -589,6 +608,7 @@ async fn test_fetch_files_glob() {
             update.source_state.as_ref(),
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await
@@ -617,6 +637,7 @@ async fn test_fetch_files_glob() {
             update5.source_state.as_ref(),
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await
@@ -672,6 +693,7 @@ async fn test_fetch_mqtt_empty() {
             None,
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             Some(listener.clone()),
         )
         .await
@@ -734,6 +756,7 @@ async fn test_fetch_mqtt_one_record() {
             None,
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             Some(listener.clone()),
         )
         .await
@@ -779,6 +802,7 @@ async fn test_fetch_container_ok() {
             None,
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             Some(listener.clone()),
         )
         .await
@@ -828,6 +852,7 @@ async fn test_fetch_container_batch_size_default() {
             None,
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await
@@ -866,6 +891,7 @@ async fn test_fetch_container_batch_size_set() {
             None,
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await
@@ -904,6 +930,7 @@ async fn test_fetch_container_batch_size_invalid_format() {
             None,
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await;
@@ -939,6 +966,7 @@ async fn test_fetch_container_has_more_no_data() {
             None,
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await
@@ -982,6 +1010,7 @@ async fn test_fetch_container_has_more_data_is_less_than_a_batch() {
             None,
             &target_path,
             &Utc::now(),
+            &HashMap::new(),
             None,
         )
         .await
@@ -1044,6 +1073,7 @@ async fn test_fetch_container_has_more_data_is_more_than_a_batch() {
                 None,
                 &target_path,
                 &Utc::now(),
+                &HashMap::new(),
                 None,
             )
             .await
@@ -1099,6 +1129,7 @@ async fn test_fetch_container_has_more_data_is_more_than_a_batch() {
                 prev_source_state.as_ref(),
                 &target_path,
                 &Utc::now(),
+                &HashMap::new(),
                 None,
             )
             .await
@@ -1153,6 +1184,7 @@ async fn test_fetch_container_has_more_data_is_more_than_a_batch() {
                 prev_source_state.as_ref(),
                 &target_path,
                 &Utc::now(),
+                &HashMap::new(),
                 None,
             )
             .await
@@ -1203,6 +1235,7 @@ async fn test_fetch_container_has_more_data_is_more_than_a_batch() {
                 prev_source_state.as_ref(),
                 &target_path,
                 &Utc::now(),
+                &HashMap::new(),
                 None,
             )
             .await
@@ -1232,7 +1265,7 @@ impl FetchTestHarness {
             None,
             None,
             None,
-            Arc::new(DatasetEnvVarServiceStaticImpl::new()),
+            Arc::new(DatasetKeyValueServiceStaticImpl::new()),
             Arc::new(RunInfoDir::new(temp_dir.path().join("run"))),
         );