+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Ingest flow panic with database api mode
+
 ## [0.191.4] - 2024-07-22
 ### Fixed
 - Parsing passwords includes specific symbols in database common crate